feat: refactor customer request handling and add security filter chain configuration
