+++ src/main/kotlin/com/crafto/crafto_backend/config/FilterChainConfiguration.kt
@@ -1,4 +1,4 @@
-package com.crafto.crafto_backend.security
+package com.crafto.crafto_backend.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -7,7 +7,7 @@ import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.web.SecurityFilterChain
 
 @Configuration
-class Configuration() {
+class FilterChainConfiguration() {
     @Bean
     fun filterChain(httpSecurity: HttpSecurity): SecurityFilterChain {
        return httpSecurity
